Centrla confidence interval for Fisher exact test
There are mulitple ways to define a p-value for the Fisher, or more
general, for 2 x 2 table test.  In order to implement
strongly consistent confidence intervals w.r.t the corresponding test,
we need to invert the corresponding test. Unfortunately, inverting the
minlike test does not necessarily lead to an interval. Thus, the default
version should be the central p-value with the corresponding interval.

This pull request adds:
- central p-value
- confidence interval based of the inversion of the central p-value
- maximum likelihood estimate of odd ratio
- tests
- some documentation