fix direction of comparison for two-tailed permutation test
previously counted number of permuted test statistics LESS extreme than the
observed statistic.  this makes the both-tail behavior consisten with right/left
tail behavior by computing the fraction of permuted statistics that are as
or more extreme.  the tests have been updated to reflect that this makes the
two-tailed test more conservative (because ties count towards h0 under this
definition).